requirements.txt
# Core Django packages
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.1

# Database and Cache
psycopg2-binary>=2.9.9
redis>=5.0.1
django-redis>=5.4.0
mssql-django>=1.3
pyodbc>=5.0.1

# Storage and AWS
boto3>=1.34.0
django-storages>=1.14.2

# Task Queue
celery>=5.3.6

# Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Monitoring and System
psutil>=5.9.8
requests>=2.31.0
python-dotenv>=1.0.0

# Security
django-environ>=0.11.2
PyJWT>=2.8.0

# Image Processing
Pillow>=10.0.0

# Development Tools
django-debug-toolbar>=4.2.0
pytest>=7.4.3
pytest-django>=4.7.0
coverage>=7.3.2 